Remove devDependencies from package.json prior to publishing
Make sure you are working on a green git before running npm publish!

# Issue
None of the dev dependencies on this project are actually needed if it is
used as a library. Therefore, remove dev dependencies prior to publishing
to NPM makes sure that this library does not polute peoples node_modules
with unneeded downloads and libraries.

For example, there is not need to install json-stable-stringify, or karma
or webpack, for using this as a third party.

Please note that running npm publish after this commit will overwrite
your package.json locally.